Cleaned up OpModes and added driveTowardsPoint opMode.
