+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/old/ArdennesNewAutoTest.java
@@ -27,44 +27,57 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.old;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.util.ElapsedTime;
 
-import java.util.ArrayList;
-
-import edu.ahs.robotics.control.Point;
-import edu.ahs.robotics.hardware.sensors.ArdennesSkyStoneDetector;
-import edu.ahs.robotics.hardware.sensors.TriggerDistanceSensor;
+import edu.ahs.robotics.control.Path;
+import edu.ahs.robotics.hardware.MecanumChassis;
 import edu.ahs.robotics.seasonrobots.Ardennes;
 import edu.ahs.robotics.util.FTCUtilities;
 import edu.ahs.robotics.util.GCodeReader;
 import edu.ahs.robotics.util.Logger;
 
 
-@Autonomous(name = "Straight Auto", group = "Linear Opmode")
+@Autonomous(name = "--- New Auto ---", group = "Linear Opmode")
 @Disabled
-public class StraightAuto extends LinearOpMode {
+public class ArdennesNewAutoTest extends LinearOpMode {
 
     @Override
     public void runOpMode() {
         FTCUtilities.setOpMode(this);
 
-        Logger logger = new Logger("pathDataStraight", "pathFollower");
+        Ardennes ardennes = new Ardennes();
+        MecanumChassis chassis = ardennes.getChassis();
+        //chassis.setPosition();
 
-        ArrayList<Point> points = new ArrayList<>();
-        points.add(new Point(0,0));
-        points.add(new Point(48,0));
+        Logger logger = new Logger("newAutoData", "pathFollower");
 
-        BaseTestAuto base = new BaseTestAuto(points, true, .2, .2);
+        Path firstMovement = new Path(GCodeReader.openFile("Intake.csv"),12,12,36);
+        Path secondMovement = new Path(GCodeReader.openFile("BackwardsToFoundation.csv"),12,12,36);
+        Path thirdMovement = new Path(GCodeReader.openFile("backUnderBridge.csv"),12,12,36);
 
         waitForStart();
 
-        base.afterStart();
+        chassis.startOdometrySystem();
 
-        stop();
+        telemetry.addLine("part 1");
+        telemetry.update();
+
+        chassis.followPath(firstMovement, true,0.2,0.15);
+
+        telemetry.addLine("part 2");
+        telemetry.update();
+
+        chassis.followPath(secondMovement, false, 0.15,0.25);
+
+        telemetry.addLine("part 3");
+        telemetry.update();
+
+        chassis.followPath(thirdMovement, true,0.1,0.1);
+
+        logger.stopWriting();
     }
 }
