+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/old/VuforiaTestOp.java
@@ -27,14 +27,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.old;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.util.Range;
+
+import edu.ahs.robotics.hardware.sensors.ArdennesSkyStoneDetector;
+import edu.ahs.robotics.util.FTCUtilities;
 
 /**
  * This file contains an example of an iterative (Non-Linear) "OpMode".
@@ -50,78 +50,62 @@ import com.qualcomm.robotcore.util.Range;
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  */
 
-@TeleOp(name="Basic: Iterative OpMode", group="Iterative Opmode")
+@TeleOp(name="Vision Test OpMode", group="Iterative Opmode")
 @Disabled
-public class OdometryTestOpMode extends OpMode
+public class VuforiaTestOp extends OpMode
 {
-    // Declare OpMode members.
-    private DcMotor x = null;
-    private DcMotor y = null;
+    //Vuforia vuforia;
+    ArdennesSkyStoneDetector detector;
+    //Bitmap vuBitmap;
+    //Bitmap croppedBitmap;
 
-    private DcMotor left, right;
+    public final int IMAGE_X = 0;
+    public final int IMAGE_Y = 500;
+    public final int IMAGE_WIDTH = 1000;
+    public final int IMAGE_HEIGHT = 200;
 
-    /*
-     * Code to run ONCE when the driver hits INIT
-     */
+
     @Override
     public void init() {
-
-        x = hardwareMap.get(DcMotor.class, "x");
-        y = hardwareMap.get(DcMotor.class, "y");
-
-        left = hardwareMap.get(DcMotor.class, "L");
-        right = hardwareMap.get(DcMotor.class, "R");
-
-        x.setDirection(DcMotor.Direction.FORWARD);
-        y.setDirection(DcMotor.Direction.FORWARD);
-
-        // Tell the driver that initialization is complete.
-        telemetry.addData("Status", "Initialized");
+        FTCUtilities.setOpMode(this);
+        //vuforia = new Vuforia(this);
+        detector = new ArdennesSkyStoneDetector(true, false);
     }
 
-    /*
-     * Code to run REPEATEDLY after the driver hits INIT, but before they hit PLAY
-     */
     @Override
     public void init_loop() {
     }
 
-    /*
-     * Code to run ONCE when the driver hits PLAY
-     */
     @Override
     public void start() {
+        detector.look();
+        //vuBitmap = vuforia.getBitmap();
+        //croppedBitmap = Bitmap.createBitmap(vuBitmap, IMAGE_X, IMAGE_Y, IMAGE_WIDTH, IMAGE_HEIGHT);
+
+        //Logger.saveImage(vuBitmap);
+        //Logger.saveImage(croppedBitmap);
+        //telemetry.addData("height", bitmap.getHeight());
+        //telemetry.addData("width", bitmap.getWidth());
+
+        //int[] pixles = new int[bitmap.getWidth()];
+
+
+//        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
+//            bitmap.getPixels();
+//            //color = Color.valueOf(bitmap.getPixel(50,50));
+//        } else {
+//            throw new Warning("Not the correct build version to work with this bitmap");
+//        }
+
 
     }
 
-    /*
-     * Code to run REPEATEDLY after the driver hits PLAY but before they hit STOP
-     */
     @Override
     public void loop() {
 
-        telemetry.addData("x",x.getCurrentPosition());
-        telemetry.addData("y",y.getCurrentPosition());
 
-        // Show the elapsed game time and wheel maxPower.
-
-        if (gamepad1.right_trigger != 0) {
-            x.setPower(-gamepad1.right_trigger);
-            y.setPower(gamepad1.right_trigger);
-        } else {
-            x.setPower(gamepad1.left_trigger);
-            y.setPower(-gamepad1.left_trigger);
-        }
-
-        left.setPower(-gamepad1.left_stick_y);
-        right.setPower(gamepad1.right_stick_y);
-
-        telemetry.update();
     }
 
-    /*
-     * Code to run ONCE after the driver hits STOP
-     */
     @Override
     public void stop() {
     }
