+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pathtests/BaseTestAuto.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.pathtests;
 
 import java.util.ArrayList;
